En weer updates!

+++ Website/Milanov/Milanov/App_Code/Products.cs
@@ -4,11 +4,11 @@
     public string Name { get; set; }
     public int Cat_id { get; set; }
     public string Cname { get; set; } // Voor het omzetten van het cat_id in een naam.
-    public double Price { get; set; }
+    public decimal Price { get; set; }
     public string Image { get; set; }
     public string Description { get; set; }
 
-    public Products(int id, string name, int cat_id, double price, string image, string description)
+    public Products(int id, string name, int cat_id, decimal price, string image, string description)
     {
         this.Id = id;
         this.Name = name;
@@ -18,7 +18,7 @@
         this.Description = description;
     }
 
-    public Products(int id, string name, string cname, double price, string image, string description)
+    public Products(int id, string name, string cname, decimal price, string image, string description)
     {
         this.Id = id;
         this.Name = name;
@@ -28,7 +28,7 @@
         this.Description = description;
     }
 
-    public Products(string name, int cat_id, double price, string image, string description)
+    public Products(string name, int cat_id, decimal price, string image, string description)
     {
         this.Name = name;
         this.Cat_id = cat_id;
